requirements.txt
streamlit
python-dotenv
pdfplumber
spacy
google-generativeai
reportlab
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
searchapi
71a0b63 (Add web_search.py)
